# resources/swagger.yaml
openapi: 3.0.0
info:
  title: Автоматизированная система управления требованиями (АСУТр)
  description: |
    API для системы управления требованиями согласно ТЗ ГОСТ 34.602-2020.
    Система предоставляет функционал для управления требованиями, их версионированием,
    согласованием, трассировкой и интеграцией со смежными системами.
  version: 1.0.0
  contact:
    name: АО "МИФИСТ"
    url: https://mifist.example.com
    email: support@mifist.example.com
servers:
  - url: https://api.asut.example.com/v1
    description: Production server
  - url: http://localhost:8000/v1
    description: Development server
...
paths:
...
  /requirements:
    get:
      tags: [Требования]
      summary: Получить список требований с фильтрами
      description: Получение списка требований с возможностью фильтрации по типу, проекту и статусу
      security:
        - bearerAuth: []
      parameters:
        - name: project_id
          in: query
          description: Фильтр по проекту
          schema:
            type: integer
            minimum: 1
          example: 1
        - name: type
          in: query
          description: Фильтр по типу требования
          schema:
            type: string
            enum: [functional, non-functional, business]
          example: "functional"
        - name: status
          in: query
          description: Фильтр по статусу требования
          schema:
            type: string
            enum: [draft, review, approved, rejected, completed]
          example: "approved"
        - name: created_by
          in: query
          description: Фильтр по создателю требования
          schema:
            type: integer
            minimum: 1
          example: 1
        - name: priority
          in: query
          description: Фильтр по приоритету
          schema:
            type: integer
            minimum: 1
            maximum: 5
            example: 1

      responses:
        '200':
          description: Список требований
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequirementListResponse'
        '400':
          description: Некорректные параметры фильтрации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Недостаточно прав для просмотра требований
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags: [Требования]
      summary: Создать требование
      description: Создание нового требования с валидацией данных
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequirementCreate'
      responses:
        '201':
          description: Требование создано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequirementFull'
  /requirements/{requirement_id}:
    get:
      tags: [Требования]
      summary: Получить требование
      description: Получение полной информации о требовании
      security:
        - bearerAuth: []
      parameters:
        - name: requirement_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Информация о требовании
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequirementFull'

    put:
      tags: [Требования]
      summary: Обновить требование
      description: Обновление требования с версионированием
      security:
        - bearerAuth: []
      parameters:
        - name: requirement_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequirementUpdate'
      responses:
        '200':
          description: Требование обновлено
        '409':
          description: Конфликт версий

    delete:
      tags: [Требования]
      summary: Удалить требование
      description: Удаление требования (мягкое удаление)
      security:
        - bearerAuth: []
      parameters:
        - name: requirement_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Требование удалено
...
  /requirements/{requirement_id}/workflow:
    get:
      tags: [Workflow]
      summary: Получить текущий статус workflow
      description: Получение текущего состояния workflow требования
      security:
        - bearerAuth: []
      parameters:
        - name: requirement_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Текущий статус workflow
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowStatus'

    post:
      tags: [Workflow]
      summary: Изменить статус workflow
      description: Изменение статуса требования в соответствии с workflow,
      security:
        - bearerAuth: []
      parameters:
        - name: requirement_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowTransition'
      responses:
        '200':
          description: Статус изменен
        '400':
          description: Недопустимый переход статуса
        '403':
          description: Нет прав для изменения статуса
  /requirements/{requirement_id}/workflow/history:
    get:
      tags: [Workflow]
      summary: История изменений workflow
      description: Получение истории изменений статуса требования
      security:
        - bearerAuth: []
      parameters:
        - name: requirement_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: История workflow
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowHistoryList'

  # ================ Согласование требований ================
  /requirements/{requirement_id}/approval:
    post:
      tags: [Workflow]
      summary: Запустить процесс согласования - 'это значит переместить в статус На проверке (Review)'
      description: Запуск процесса согласования требования
      security:
        - bearerAuth: []
      parameters:
        - name: requirement_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Процесс согласования запущен
        '400':
          description: Недопустимый переход статуса
        '403':
          description: Нет прав для изменения статуса
  /requirements/{requirement_id}/approve:
    post:
      tags: [Workflow]
      summary: Утвердить требование
      description: Утверждение требования уполномоченным лицом
      security:
        - bearerAuth: []
      parameters:
        - name: requirement_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                comment:
                  type: string
      responses:
        '200':
          description: Требование утверждено
        '403':
          description: Нет прав на утверждение
  /requirements/{requirement_id}/reject:
    post:
      tags: [Workflow]
      summary: Отклонить требование
      description: Отклонение требования с комментарием
      security:
        - bearerAuth: []
      parameters:
        - name: requirement_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - comment
                - reason
              properties:
                comment:
                  type: string
                reason:
                  type: string
                  enum: [incorrect, incomplete, duplicate, outdated]
      responses:
        '200':
          description: Требование отклонено
        '403':
          description: Нет прав на отклонение

  # ================ Обсуждения и комментарии ================
  # /requirements/{requirement_id}/discussions:
  #   get:
  #     tags: [Обсуждения]
  #     summary: Получить обсуждения требования
  #     description: Получение всех обсуждений по требованию
  #     security:
  #       - bearerAuth: []
  #     parameters:
  #       - name: requirement_id
  #         in: path
  #         required: true
  #         schema:
  #           type: integer
  #     responses:
  #       '200':
  #         description: Список обсуждений
  #         content:
  #           application/json:
  #             schema:
  #               $ref: '#/components/schemas/DiscussionList'

    # post:
    #   tags: [Обсуждения]
    #   summary: Создать обсуждение
    #   description: Создание нового обсуждения по требованию
    #   security:
    #     - bearerAuth: []
    #   parameters:
    #     - name: requirement_id
    #       in: path
    #       required: true
    #       schema:
    #         type: integer
    #   requestBody:
    #     required: true
    #     content:
    #       application/json:
    #         schema:
    #           $ref: '#/components/schemas/DiscussionCreate'
    #   responses:
    #     '201':
    #       description: Обсуждение создано

  # /discussions/{discussion_id}/comments:
  #   get:
  #     tags: [Обсуждения]
  #     summary: Получить комментарии обсуждения
  #     description: Получение всех комментариев в обсуждении
  #     security:
  #       - bearerAuth: []
  #     parameters:
  #       - name: discussion_id
  #         in: path
  #         required: true
  #         schema:
  #           type: integer
  #     responses:
  #       '200':
  #         description: Список комментариев
  #         content:
  #           application/json:
  #             schema:
  #               $ref: '#/components/schemas/CommentList'

  #   post:
  #     tags: [Обсуждения]
  #     summary: Добавить комментарий
  #     description: Добавление комментария к обсуждению
  #     security:
  #       - bearerAuth: []
  #     parameters:
  #       - name: discussion_id
  #         in: path
  #         required: true
  #         schema:
  #           type: integer
  #     requestBody:
  #       required: true
  #       content:
  #         application/json:
  #           schema:
  #             $ref: '#/components/schemas/CommentCreate'
  #     responses:
  #       '201':
  #         description: Комментарий добавлен

  # ================ Зависимости требований ================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
        timestamp:
          type: string
          format: date-time
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        pages:
          type: integer

    # ================ Пользователи ================
...
    Requirement:
      type: object
      properties:
        id:
          type: integer
        project_id:
          type: integer
        name:
          type: string
        type:
          type: string
        created_at:
          type: string
          format: date-time
        created_by_user_id:
          type: integer
        path:
          type: string
        depth:
          type: integer
        requirement_group_id:
          type: integer
    RequirementFull:
      allOf:
        - $ref: '#/components/schemas/Requirement'
        - type: object
          properties:
            content:
              $ref: '#/components/schemas/RequirementContent'
            workflow:
              $ref: '#/components/schemas/RequirementWorkflow'
            dependencies:
              type: array
              items:
                $ref: '#/components/schemas/Dependency'
    RequirementCreate:
      type: object
      required:
        - project_id
        - name
        - description_text
      properties:
        project_id:
          type: integer
        name:
          type: string
          maxLength: 255
        type:
          type: string
        parent_id:
          type: integer
        content:
          $ref: '#/components/schemas/RequirementContentCreate'
    RequirementUpdate:
      type: object
      properties:
        name:
          type: string
          maxLength: 255
        type:
          type: string
    RequirementListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/RequirementFull'
        pagination:
          $ref: '#/components/schemas/Pagination'

    # ================ Содержимое требований ================
    RequirementContent:
      type: object
      properties:
        id:
          type: integer
        development_basis:
          type: string
        development_purpose:
          type: string
        description_text:
          type: string
        acceptance_criteria:
          type: string
        document_requires:
          type: string
        requirement_id:
          type: integer
    RequirementContentCreate:
      type: object
      properties:
        development_basis:
          type: string
        development_purpose:
          type: string
        description_text:
          type: string
        acceptance_criteria:
          type: string
        document_requires:
          type: string
...
    RequirementWorkflow:
      type: object
      properties:
        id:
          type: integer
        requirement_content_id:
          type: integer
        priority:
          type: integer
          minimum: 1
          maximum: 5
        status:
          type: string
          enum: [draft, review, approved, rejected, implemented, verified, archived]
    WorkflowStatus:
      allOf:
        - $ref: '#/components/schemas/RequirementWorkflow'
        - type: object
          properties:
            requirement_name:
              type: string
            can_transition_to:
              type: array
              items:
                type: string
    WorkflowTransition:
      type: object
      required:
        - new_status
      properties:
        new_status:
          type: string
          enum: [draft, review, approved, rejected, implemented, verified, archived]
        comment:
          type: string
    WorkflowHistoryList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/WorkflowHistoryItem'
    WorkflowHistoryItem:
      type: object
      properties:
        id:
          type: integer
        prev_req_workflow_id:
          type: integer
        created_by_user_id:
          type: integer
        created_at:
          type: string
          format: date-time
        is_active:
          type: boolean
        workflow:
          $ref: '#/components/schemas/RequirementWorkflow'

    # ================ Обсуждения ================
    # Comment:
    #   type: object
    #   properties:
    #     id:
    #       type: integer
    #     discussion_id:
    #       type: integer
    #     created_by_user_id:
    #       type: integer
    #     created_at:
    #       type: string
    #       format: date-time
    #     content:
    #       type: string

    # CommentCreate:
    #   type: object
    #   required:
    #     - content
    #   properties:
    #     content:
    #       type: string

    # CommentList:
    #   type: object
    #   properties:
    #     data:
    #       type: array
    #       items:
    #         $ref: '#/components/schemas/Comment'

    # ================ Зависимости ================
    Dependency:
      type: object
      properties:
        id:
          type: integer
        target_requirement_id:
          type: integer
        source_requirement_id:
          type: integer
        type:
          type: string
          enum: [blocks, depends_on, related_to, duplicates, refines]
        strength:
          type: integer
          minimum: 1
          maximum: 5
        description:
          type: string
        source_requirement_group_id:
          type: integer
        target_requirement_group_id:
          type: integer
...